Stubbing some repo files
Need some modules to handle getting and putting data into the data
store. The concept of Repo's are, again, taken from Ecto, and I don't
know if it is needed when the Riak-module will take care of the
connection to the cluster, but I think I will go down this route for
now.

Perhaps there are other CRDT stores, than Riak KV, that I could
implement adapters for in the future.