Generate random key for new entries
The function generating random keys can be overwritten and receive the
changeset so the implementor can decide how to specify the key. The
default option rely on the UUID Elixir package. This might get replaced
by a home grown one to reduce the number of dependencies at a later
stage.