Added a context to the flags
For now I just use the key, I don't know if this has any bad
implications, but at least it seems to get rid of the
`:context_required` exception when initializing a flag as
disabled. Later I might just treat the absence of a flag as `false`, and
its value as its value, true or false, if it has been set in the past.