Implemented base functionality of casting
Allows for accepting data from the outside, such as HTTP POSTs. This
code has been (like great parts of this project) lifted from Ecto; It
does away with the required/optional values and goes with a list
of *allowed* fields instead; this is how Ecto will do in the future.